Remove pnp sign IO
